updating require references
