+++ centralization/services.js
@@ -1,10 +1,10 @@
-var aws = require("./aws.js");
-var auth = require('../authentication/services.js');
-
-function registerCentralizationServices (app) {
-  app.all('/centralization/*', auth.checkJwt);
-  app.post('/centralization/upload', aws.uploadFile);
-  app.get('/centralization/submitted-files', aws.getSubmittedFiles);
-}
-
-exports.registerCentralizationServices = registerCentralizationServices;
+// var aws = require("./aws.js");
+// var auth = require('../authentication/services.js');
+//
+// function registerCentralizationServices (app) {
+//   app.all('/centralization/*', auth.checkJwt);
+//   app.post('/centralization/upload', aws.uploadFile);
+//   app.get('/centralization/submitted-files', aws.getSubmittedFiles);
+// }
+//
+// exports.registerCentralizationServices = registerCentralizationServices;
